vooooiiillllaaaaa

+++ c03/ex02/ft_strcat.c
@@ -6,14 +6,16 @@
 /*   By: hucherea <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/14 08:47:27 by hucherea          #+#    #+#             */
-/*   Updated: 2024/02/14 08:49:52 by hucherea         ###   ########.fr       */
+/*   Updated: 2024/02/14 14:20:10 by hucherea         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <aio.h>
+
 char	*ft_strcat(char *dest, char *src)
 {
-	int	i_dest;
-	int	i_src;
+	size_t	i_dest;
+	size_t	i_src;
 
 	i_dest = 0;
 	i_src = 0;
